Shuffle files around + add DEV_SERVER env variable for static files
