+++ .env
@@ -2,3 +2,4 @@ HOST=127.0.0.1
 PORT=3001
 CACHE_DIR=/tmp/rustatouille
 DB_CONNECTION=sqlite:///tmp/rustatouille.db?mode=rwc
+DEV_SERVER=true
