+++ src/controllers/mod.rs
@@ -1,0 +1,2 @@
+pub mod admin;
+pub mod r#static;
