+++ src/db/models/comments.rs
@@ -1,0 +1,8 @@
+use chrono::NaiveDateTime;
+
+#[allow(dead_code)]
+#[derive(sqlx::FromRow)]
+pub struct Comment {
+    date: NaiveDateTime,
+    description: String,
+}
