+++ src/db/models/mod.rs
@@ -1,0 +1,3 @@
+pub mod comments;
+pub mod interventions;
+pub mod services;
