+++ src/main.rs
@@ -4,14 +4,18 @@ use axum::{
     Extension, Router,
 };
 use sqlx::AnyConnection;
-use std::{env, net::Ipv4Addr, path::PathBuf, sync::Arc};
+use std::{
+    env,
+    net::Ipv4Addr,
+    path::{Path, PathBuf},
+    sync::Arc,
+};
 use std::{fs, net::SocketAddr};
 use tokio::sync::Mutex;
 use tracing as log;
 
-mod admin;
+mod controllers;
 mod db;
-mod r#static;
 
 pub(crate) struct AppConfig {
     /// which port the app is listening on
@@ -25,6 +29,9 @@ pub(crate) struct AppConfig {
 
     /// Path to the sqlite file
     db_connection_string: String,
+
+    /// Should the server also respond to static queries, in dev mode?
+    dev_server: bool,
 }
 
 pub(crate) struct AppContext {
@@ -60,33 +67,42 @@ fn parse_app_config() -> anyhow::Result<AppConfig> {
         .context("DB_CONNECTION doesn't designate an utf8 path")?
         .to_owned();
 
+    let dev_server = env::var("DEV_SERVER")
+        .context("missing DEV_SERVER env")?
+        .to_lowercase();
+    let dev_server = ["true", "yes", "y"].iter().any(|v| dev_server == *v);
+
     Ok(AppConfig {
         port,
         interface_ipv4,
         cache_dir,
         db_connection_string,
+        dev_server,
     })
 }
 
-fn copy_static_files_to_cache_dir(ctx: &AppContext) -> anyhow::Result<()> {
+/// Copy the static files to the cache directory.
+///
+/// TODO: should also do it on change on disk, in the DEV_SERVER mode?
+fn copy_static_files_to_cache_dir(cache_dir: &Path) -> anyhow::Result<()> {
     // Copy the style.
     let style = include_str!("./view/style.css");
-    fs::write(ctx.config.cache_dir.join("style.css"), style)?;
+    fs::write(cache_dir.join("style.css"), style)?;
 
     let style = include_str!("./view/admin.css");
-    fs::write(ctx.config.cache_dir.join("admin.css"), style)?;
+    fs::write(cache_dir.join("admin.css"), style)?;
 
     Ok(())
 }
 
 async fn real_main() -> anyhow::Result<()> {
-    // initialize tracing
+    // Initialize tracing.
     tracing_subscriber::fmt::init();
 
+    // Parse the configuration.
     let config = parse_app_config()?;
 
-    let listen_addr = SocketAddr::from((config.interface_ipv4, config.port));
-
+    // Start the database.
     let conn = db::open(&config.db_connection_string).await?;
 
     let ctx = Arc::new(AppContext {
@@ -94,21 +110,36 @@ async fn real_main() -> anyhow::Result<()> {
         db_connection: Mutex::new(conn),
     });
 
-    copy_static_files_to_cache_dir(&ctx)?;
+    // Generate the full web site initially.
+    copy_static_files_to_cache_dir(&ctx.config.cache_dir)?;
 
-    // build our application with a route
-    let app = Router::new()
-        .route("/admin", get(admin::index))
-        .route("/admin/service/new", get(admin::create_service_form))
+    // Configure and start the web server.
+    let mut app = Router::new()
+        .route("/admin", get(controllers::admin::index))
+        .route(
+            "/admin/service/new",
+            get(controllers::admin::create_service_form),
+        )
         .route(
             "/admin/intervention/new",
-            get(admin::create_intervention_form),
+            get(controllers::admin::create_intervention_form),
         )
-        .route("/admin/api/service", post(admin::create_service))
-        .route("/admin/api/intervention", post(admin::create_intervention))
-        .route("/*path", get(r#static::get)) // catch-all
-        .layer(Extension(ctx));
+        .route(
+            "/admin/api/service",
+            post(controllers::admin::create_service),
+        )
+        .route(
+            "/admin/api/intervention",
+            post(controllers::admin::create_intervention),
+        );
 
+    if ctx.config.dev_server {
+        app = app.route("/*path", get(controllers::r#static::get)); // catch-all
+    }
+
+    app = app.layer(Extension(ctx.clone()));
+
+    let listen_addr = SocketAddr::from((ctx.config.interface_ipv4, ctx.config.port));
     log::debug!("listening on {}", listen_addr);
 
     // This, in fact, will never return.
